btc-bot/requirements.txt
python-binance==1.0.19
yfinance>=1.2.0
pandas==2.2.2
requests==2.32.3
python-telegram-bot==21.3
ntscraper==0.3.9
schedule==1.2.2
flask==3.0.3
